requirements.txt
Pillow==8.4.0
pandas==1.2.3
openpyxl==3.1.2